+++ smart-watering-backend/src/main/java/com/iot/smartwatering/smart_watering_backend/repository/ScheduleRepository.java
@@ -1,14 +1,17 @@
 package com.iot.smartwatering.smart_watering_backend.repository;
 
-import com.iot.smartwatering.smart_watering_backend.entity.Schedule;
+import java.util.List;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
+import com.iot.smartwatering.smart_watering_backend.entity.Schedule;
 
 @Repository
-public interface ScheduleRepository extends JpaRepository<Schedule, Integer> {
-    List<Schedule> findByZone_ZoneId(Integer zoneId);
+public interface ScheduleRepository extends JpaRepository<Schedule, Long> {
+    List<Schedule> findByZone_ZoneId(Long zoneId);
+
     List<Schedule> findByActiveTrue();
-    List<Schedule> findByZone_ZoneIdAndActiveTrue(Integer zoneId);
+
+    List<Schedule> findByZone_ZoneIdAndActiveTrue(Long zoneId);
 }
